requirements.txt
streamlit==1.45.1
pandas==2.2.3
plotly==5.24.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
pillow==11.3.0
requests==2.32.3
openai>=1.30.0
